requirements.txt
# Python3

PyExecJS

